requirements.txt
astronomer-cosmos
apache-airflow-providers-snowflake
/tmp/astro_provider_snowflake-0.0.0-py3-none-any.whl